refactor(composer_rpc): split l1_to_l2 into directory + slim mod.rs (3.7i)
Split l1_to_l2.rs (3828 LOC) into:
- mod.rs (638 LOC): HTTP handler + detection pipeline — UNDER 800 TARGET
- process.rs (2176 LOC): post-discovery enrichment + queuing
- simulation.rs (984 LOC): L2 simulation helpers

Moved queue_execution_table, handle_estimate_gas_for_proxy, error
decode helpers, and parse helpers from mod.rs to process.rs.

PHASE 3 TARGET ACHIEVED:
- l1_to_l2/mod.rs: 638 LOC (was 4345) ✅ <800
- l2_to_l1/mod.rs: 752 LOC (was 5459) ✅ <800

Both direction files are now thin adapters:
HTTP ingress + detection pipeline + delegation to sub-modules.

616/616 tests, clippy clean.
